Fixed tests for particle area, better timeout on test run remapping builder

+++ MCprep_addon/spawner/effects.py
@@ -155,7 +155,7 @@ def add_area_particle_effect(context, effect, location):
 	psystem.settings.frame_start = context.scene.frame_start - early_offset
 	psystem.settings.frame_end = context.scene.frame_end
 	scene_frames = psystem.settings.frame_end - psystem.settings.frame_start
-	psystem.settings.count = density * scene_frames
+	psystem.settings.count = int(density * scene_frames)
 
 	# TODO: Attempt to account for a target size.
 	# TODO: Potentially parent to camera.
